Possible fix for Alliance Overhaul exception
Activeweapon is not a member of this, m, n, or d.   This.weapon.unit /
this has an activeweapon member but in this context it's set to -1,
which would probably cause an out of range exception too.

But each weapon has the .isactive member which, at least in this case,
is true.  So this seems like a safe way to select a valid active weapon
for Alliance Overhaul-equipped ARC-170s.