Adding compiled javascript after all
I was holding off on checking in the compiled version, so that the
XWS-Bench repo could import my changes without immediately changing the
hosted code itself.  But I eventually decided to check in the compiled,
minified code so that people can test my changes for themselves, via my
cloned repo.