Continuing work on adjusting map to Imperial measurements
This should update all maps and starting zones to Imperial (36" x 36")
measurements.
The changes to the many, many "max=Math.max(...)" statements are purely
aesthetic as the math works the same either way, but looks cleaner and
is more modular.

Of note: positional initialization of units actually lives in team.js.
I've added a comment to mark this for future reference.