Fix for Poe + Advanced Optics + Jan Ors, plus some minor changes
1) Fixed Poe, Advanced Optics, Jan Ors interaction.  Previously Jan
"wrapped" addfocustoken(), which meant that after a focus token was
added, she then either did nothing or removed a focus and added an
evasive.  But the *actual* way she should work is either A) add a focus,
or B) prevent a focus from being added, and instead add an evasive.  I
have implemented this functionality by making addfocustoken() a No-Op,
and having Jan Ors choose whether a call to addfocustoken() actually
adds a focus token or instead adds an evade token.  Note: if any other
upgrade card or pilot overrides Unit.prototype.addfocustoken in the
future, either it or Jan Ors will be broken.

2) Fixed a problem with Calculation that is normally invisible, but is
exposed with Garven Dreis.  The problem caused dozens of focus tokens to
be added to Garven's teammates when Garven + R2-D6 + Calculation was
used.  The root problem is that tohitproba() repeatedly calls
modifyattackroll() when figuring out the best attack probability, so any
wrapped function within wrapping functions on modifyattackroll() will be
called many times.  I suspect this is behind the Biggs stress-storm
issue as well.  Solution was to replaces this.removefocustoken() with a
bare this.focus--; the wrap_after will only ever be called during
tohitproba() so there's no need for the more UI-oriented function call.

3) Documented some points that gave me problems while debugging #1 and
#2

4) Fixed a tiny typo in IAUnit during the action choice loop: also added
a check that should have pilots select "Evade" actions if they already
have a Focus token.