Check to prevent R4-D6 triggering on its ship's target; also Corran Horn work
1) Two R4-D6s attacking each other (especially problematic on Biggs +
R4-D6 mirror matches) currently will trigger twice: once from the
defender modifying the attack dice, and once from the attacker
(erroneously) modifying the attack dice again.  Ensuring that it only
fires when the defender is using it prevents this.

2) Two Corran Horns can't both take their end-of-phase 2nd attacks
because of (apparently) a lack of sequential execution; basically, the
first instance gets overwritten by the second, and then the phase ends.
But adding ".bind(this)" to the second attack function should make the
context of that function ironclad.  Work is continuing on this issue.