First step on Thweek fix
Moved Thweek selection from resolveactionselection() to a selectunit()
call in order to provide the ability to cancel the selection.

Note: this selectunit call does not block so the next phase begins as
soon as the selection UI updates are applied.  I'm looking into this;
whatever the fix is, it will probably fix the "dueling Corran Horn" as
well.