Fix for Ionized + Shaken Pilot
As expected, the internal handling of Ionized generates a "dial" of a
single white forward-1 move, rather than giving the affected ship a
mandatory white-forward-1 maneuver directly.  Rather than figure out how
to pass a maneuver instead of a dial, I added a check for the specific
case.

N.B. This interaction is not covered in any FAQ, but there are several
rules threads on the FFG forums supporting the reading that Ionization
does not assign a maneuver or dial, so its mandatory 1WF move is not
affected by Shaken Pilot.