Fix for broken Maarek Stele effect
While researching "$.Deferred"-based locking and callbacks, I found that
Maarek Stele's critical-choosing effect was not working.  It looks as
though a change was made to "activeunit" such that it does not reflect
the currently attacking unit by the time execution reaches this section
of code.  So I added an "attackunit" global, the logical complement to
the existing "targetunit".