Final fix for Rigged Cargo Chute (and, by extension, all Obstacle collisions)
I found a couple of irregularities in obstacle collision checking:

1) A "template" wouldn't be checked if the current ship had previously
hit the currently-checking object in a previous turn (this is at odds
with the obstacle rules)
2) "Template" paths are generated center-to-center rather than
front-nubs to back-nubs.  In practice this means that a ship whose back
half was on an obstacle could be counted as having had its template pass
over that obstacle, even if a real template never would have come near
it.  I've added some code to exclude path points that lie within the
start- or end-points' base areas from these template checks.  I will
probably also do a once-over on the mine collision code as well, later.