Fix for issue #20 - empty list -> weird behavior
Pressing (1) or (2) in squad creation screen without adding any ships
causes the game to try to create and populate a squad with no actual
name, no ships, and no points.  It looks like the original dev attempted
to protect against *some* undefined-variable issues, but not this
glaring deficiency.

I've wrapped the list-entry population section in an if-statement that
checks if the list string is empty or not.  This seems to avoid creating
an empty list, which should avoid the subsequent odd issue with deleting
an actual list instead of the empty list as well.