Fix for issue #15 - hang when AI Striker has to take its AA move off the board
The IAUnits maneuver selection code was returning a "-1" value into a
section of code that was expecting a valid array index, leading to a
hang later on in the action resolution code.  Since this only occurs
when an AI ship has no option but to fly off the table using one of its
AA moves, I didn't see a reason to make the fix very complex.

I have verified that this code still allows the AI ship to choose a
different Adaptive Ailerons maneuver if that move will allow it to
safely return to the map.