Final part of Kylo Ren fix
It turns out that, for pilots who are on multiple ships, the internal
code needs to know that they are "ambiguous" and what each subsequent
version's edition is.  This allows it to correctly compose and parse its
own versions of the List Juggler strings which the list creation page
puts out.