Fix for issue #39 (and incorrect Boba Fett crew action)
During testing simple fix for issue 39, I found that Boba Fett was not
removing Title or Mod cards as he should.  This generally is not a
problem because most Title-granted effects are not removed when the
Title card itself is.  However, in the Heavy Scyk case the effect is the
same as for Hull Upgrade: the additional hull point is removed:

https://community.fantasyflightgames.com/topic/232160-boba-fett-and-hull-upgrade/?page=2&tab=comments#comment-2463997