2018/01/29 - 2018/02/04 bug smasher pass (#38)
* Integrated Astromech triggering multiple times on one roll (Issue #36)

Integrated Astromech did not check its activation state within the
callback that actually determines wether IA can discard a damage or not,
only within the initial setup of an attack roll.  Adding another active
state check within the callback itself prevents IA from triggering more
than once in a given attack resolution.

* Fix for issue #39 (and incorrect Boba Fett crew action)

During testing simple fix for issue 39, I found that Boba Fett was not
removing Title or Mod cards as he should.  This generally is not a
problem because most Title-granted effects are not removed when the
Title card itself is.  However, in the Heavy Scyk case the effect is the
same as for Hull Upgrade: the additional hull point is removed:

https://community.fantasyflightgames.com/topic/232160-boba-fett-and-hull-upgrade/?page=2&tab=comments#comment-2463997

* Fix for certain Upgrade cards getting installed twice.

While working on other bug fixes I was reminded of an issue I'd seen
with some upgrades, mainly those like Engine Upgrade which add an action
to the equipping ship's action bar.  These upgrades were providing not
one but two action icons, and when removed (as with Burnout Slam) would
leave one active icon in place.

It appears that the team construction code attempts to handle all
upgrades via the Upgrade class install function, but also wants to run
specific cards' install() functions when they are available.  However,
the check that was in place did not correctly differentiate between
cards that inherit the superclass' default install() function and those
that have special install() functions.  Thus, some Upgrade cards were
having the default install function run on them twice.

We really only need to run the subclass install() function if it differs
from the default; otherwise, one pass of Upgrade.prototype.install() is
sufficient.