Final futzing with Hound's Tooth / Docking / Deploy
There is a major issue with the "create squad from row" functionality.
It only superficially creates a squadron when the user selects a saved
squad for editing.  I have attempted to normalize this functionality to
degree so that editing a squad with Hound's Tooth and Nashtah Pup does
not e.g. cause extraneous ship silhouettes to appear during the game,
but there is still much work to be done in this area.