Fix for issue #47 (Multiple Boba Fett procs)
Implemented a rudimentary semaphore that appears to correctly stop more
than one proccing of Boba Fett on an attack with multiple crits, yet
allows skipping one crit and proccing on another within the same attack.
At any rate, the low-frequency issue I was seeing now seems to be a
zero-frequency.