Fix for VCX-100 exception when using Torpedoes without Ghost/Phantom
This is probably a fairly uncommon use case, but I found that flying
Lothal Rebels *without* Ghost and Phantom and with Torpedoes equpped
caused an exception fairly frequently.

It turns out that the code to check if another ship is in a weapon's
auxiliary firing arc did not know how to handle a torpedo with an
auxiliary arc, and was defaulting to using the ship's primary weapon.
This doesn't cause a bug when Ghost and Phantom are used because then
the primary weapon is set to Bilaser and has an auxiliary arc - doing
the wrong thing, but getting the right result.

The fix was to make sure to explicitly call the getauxiliarysector()
function of any non-laser, non-normally-auxiliary weapon.  Currently
only VCX-100s and Moralo Eval are affected, but I've tested this with
both sets and it's working correctly.