Refinement of docked checks, AI + this.init compatibility
I found a couple of subtle targetting and IAunit issues while
double-checking the deploy code.

1) Most paths successfully prevented selecting or attacking docked
ships, but not all: closed a couple of loopholes that let e.g.
autoblaster turret target the invisible "docked" Nashtah Pup Pilot in
the top-right corner of the map.

2) IAUnits2.js contains some experimental code that needs attributes
added to this.init.  On the other hand, several upgrades explicitly call
.init() on ships in order to copy pilot abilities around.  Rather than
break either, I combined them.  This *should* fix a few bugs when
fighting AI units without invalidating any of the IAUnits2 code, but
since we're not using that code I'd be fine with just abandoning it
completely as well.