Fix for issue #49: AI Ghost always deploys the Phantom/Phantom II immediately
The current simplistic AI action decision tree simply takes the first
action of any set of actions if that action is not a Critical card
action, an Evade action, or a Focus action.

I've added a very simple mechanism to arbitrarily specify an AI-specific
decision function on any action-providing upgrade or pilot (I think),
and to call that function if it exists.  I've named it aiactivate to
match with the convention for all addattackmodifier specs, but it could
easily be named something else if that causes confusion.

Right now only the Ghost uses this function, and it simply chooses to
deploy the docked shuttle if the Ghost's hitpoint total (hull + shield)
drops below 5, so it's super-arbitrary.  The function call does not
provide any parameters either; all the AI function can know about is its
own state, that of its attached ship, and any globals.  Hopefully that
is enough to build a framework on, though.