Fix for xws-bench/issues#15
Bugs:

1) Dash Rendar pilot could not attack on a turn it landed on an
obstacle, even if it subsequently barrel-rolled or boosted off.

2) Dash Rendar + Countermeasures + Proton Rockets resulted in 4 attack
dice, not 5 as expected.

Fixes:

1) If a resolved action move does not hit any obstacles, reset
ocollisions.overlap back to -1.  NOTE: this does not remove the
ocollision instance, or re-run ocollision checks, so if any effect
relies on the ocollision instance to see if a ship has collided with any
obstacles in the past turn, that effect should continue working as
expected.

2) Changed prockets to use the attached unit's .getagility() function,
which returns the correct value.  Suggest that all abilities that rely
on ship agility use this function going forward, as it appears to be the
original dev's intention.