Added bomb priority so AI can use action bombs
While this may look like it prioritizes bombs over almost every other
action, this is simply in order to have the AI pilot check if enemies
are in positions where bombing *may* be useful.  If the AI checks but
decides that bombing is not justified, it will continue on to Cloak,
Focus, Target, Evade, etc.