Added aiactivate for Black Market Slicer Tools as a test
This is the first non-bomb-style ILLICIT to get an aiactivate()
function.
It seems to be working well without causing undue slowdowns.
Equipped ships will choose to use the BMST depending on a few criteria,
and assuming that a Range 1 attack is almost always better than BMST
unless the target is high-DEF + low remaining hull.