Naive approach to incentivizing green manuevers on multi-action ships with stress
Currently the AI almost never clears stress.  This change increases the
AI's affinity for green manuevers when the ship has multiple actions, or
additional actions, with the affinity increasing as stress goes up.
This can always be tweaked further; one idea would be to make the stress
weight exponential but remove the default action count.