POC working but needs tons of locking
With $(document).on(), each instance of Jan Ors needs to know if another
instance is in play or not, because what it does for enemies changes
depending on that information.
The *best* option would be to have only friendlies trigger their team's
version of Jan, which has its own issues - what is the top-level
team-related element that can be triggered?  How do we "wrap" only one
team - if we could do this, the event system wouldn't be needed!

A global lock / semaphore / is the easiest to implement, but the most
disgusting - and anyhow, that could be done just as easily with the
wrap_before functionality.