Final working version of Jan Ors that covers all possibilities
There are a number of combinations that have to be checked here,
including some variety of these states:

1) Jan Ors is on "this" team
2) Jan Ors is active
3) Jan Ors is within 3
4) The current ship is / is not already choosing whether to use Jan Ors
5) Jan Ors has died

The new collection of function redefinitions, event triggers, and
locks/semaphores is necessary to ensure that every single possibility,
including when one team has Jan but the other doesn't *and* when both
teams have their own Jan, is handled correctly.

In the end this wasn't a great test of the event listener paradigm but
it shows how a listener/trigger pair can be used with team numbers to
only trigger the appropriate listener for a team.  I would like to make
this an integral part of some other functions (such as endmaneuver) so
that Snap Shot and Corran Horn's abilities can be handled correctly.