Atomic Snap Shot that correctly blocks other activations... but only for IAUnits
This is very frustrating.  It looks like Player-controlled ships have
additional layers of UI interaction that prevent this form of
interrupting donoaction from completing; on the other hand, this works
perfectly for IAunits.

Apparently, doattack() for regular Units involves several layers of
selection - one to choose a weapon system, one to select a target for
that attack from the valid targets in range for the weapon selected, and
then a third that enqueues and *should* resolve an action - but which
breaks the chain of actions.  Somehow we either never resolve the
attack, or never even get to the point where the user should be
presented with the attack UI in order to click and resolve the attack.

I believe that it should be possible to make this work for IAUnits *and*
Player Units using a modified version of the IAUnits code.  However,
this is taking some time to nail down.