New version of Snap Shot renders this problematic section unnecessary
I had added this new end-of-activation unlock section specifically to
handle the ending of Snap Shot actions, but with the prior hotfix and
the addition of new lock-handling mechanics in Snap Shot, this section
is no longer necessary.

I should note that this section would have needed to be adjusted anyhow,
as it seems to interfere with Attack-phase additional actions and
attacks.  Removing it is the cleaner fix.