First pass at Adaptive Ailerons AI update
Currently the AI is incredibly bad at utilizing A.A.  The first turn for
AI TIE Strikers is usually "turn left, fly off the board". I was forced
to deal with this while testing the last two hotfixes.  The main issue
is that the AI has to plan two maneuvers without knowing what either of
them are, so it cannot coordinate the A.A. move with its planning phase.

This approach first chooses the "best" maneuver from the A.A. mandatory
moves, and then temporarily places the AI ship at the new position to
compute the regular maneuver as per usual.   It's not the closest to TT
rules but it does have the advantage of being simple to implement.