Updates to some older, bugged ships / cards + endsetupphase()
Running back through old XWS-Bench bugs got me into Hyperspace Comm
Scanner, Kylo Ren's Shuttle, and more Glitterstim work.

1) In order for abilities that proc at the end of the Setup phase to
work correctly, we have to instigate our own action locking (as the base
engine does not begin handling actions correctly until the Activation
phase).  To that end I added a rudimentary version of the existing
action queue to endsetupphase() within xwings.js.  I also added more
.trigger / .on event handling.
Note: in order to allow the Setup phase to end while still handling all
possible abilities without overwriting them, I made endsetupphase either
exit to the base Document event handler loop (if there are actions to
handle before continuing) or go to the next phase as normal.
Unfortunately the cleanest way to do this was to run the SETUP_PHASE
case of nextphase() twice.  It's ugly, but it should be solid.

At any rate, it is now possible to handle multiple abilities that
trigger at the end of the setup phase, or even (as is the case with HCS)
multiple copies of the same ability if necessary.

2) Re-wrote Hyperspace Comm Scanner to work properly as per the card
text.
A) the user gets to choose whether the attached ship gets PS 0, 6, or 12
during setup.  If any additional copies of HCS are attached, their
PS-switching code is disabled.
B) Any ships that would be set up "after" the HCS ship (that is, with an
equal or higher PS than the selected value) gets one action to add a
Focus or Evade token per equipped copy of HCS.
C) At the end of setup, the attached ship's PS goes back to whatever it
should be (accounting for VI and Adaptability correctly).

3) Made the aiactivate() function for Glitterstim more intelligent; it
will now only let the AI use Glitterstim if A) it hasn't already used
one this round, and B) enemy ships are nearby.

4) Closed a hole in the Kylo Ren's Shuttle title which allowed single
enemies without any allies nearby to avoid getting the Stress token.
It's still not quite correct but it assigns a token no matter what now.