Intensity 90% finished
AI usage is not great, and it's possible to proc Intensity's Focus/Evade
benefit twice in one turn on ships with enough actions to use Boost and
Barrel Roll in the same Activation, but it's pretty much correct and the
AI does flip the card appropriately *most* of the time.  Good enough for
now.