Added 'Double Edge', also keyboard shortcut info on adding/removing Weapons Disabled tokens
Double edge is working properly with all manner of weapons.  I highly
recommend deactivating weapons during additional attacks as a method of
avoiding complex this.weapons filtering; as long as you reactivate the
weapon at the end of the combat phase this works fine.

I will point out, currently there is an end-of-activation issue when
there is no other attacker after Double Edge and Double Edge uses a
dual-attack weapon (Twin Laser Turret, Cluster Missiles) as the
secondary attack that misses.  This leads to the action queue exhausting
prior to the user getting to interact with the combat dial, which clears
the dial and halts execution (or more properly, returns execution to the
event handler loop without any way to signal that the user wants to fire
or skip the added attack).  I think this can be fixed by adding a null
action somewhere, probably within the addattack code.  I'll check it
out.