Updated Shara Bey to work correctly
Previously Shara's wrapping A) did not work during the check to see if a
secondary weapon's Target Lock requirement was met, and B) would not
work for both teams if each team had a Shara.

I moved the canusetarget() wrapper to the start of the combat phase, and
used the standard TEAM[this.team].has<ability host> /
$(document).trigger(<phase>+<team ID>,...) paradigm to let each team
have its own Shara handlers.