First pass at Rebel Nym
Lots of updates here and there to facilitate post-activation actions
regarding bombs:

1) Added another lightweight action queue handler in xwing.js,
specifically for bombs.  Also instrumented the bomb explosion section
with event triggers to notify e.g. Nym that bombs want to explode.

2) instrumented units.js for mine collisions with pre- and
post-explosion event triggers.  This area is simple but could probably
be cleaned up.

3) Added some Unit.prototype functions relating to click handling into
the Bomb prototype so that bombs can be selected, and implemented the
basic triggers that Nym consumes.

4) Added handling for bombs and mines in the Captain Nym init() section.

I've tested this with bombs and mines, and confirmed that mirror matches
work correctly.  There should probably also be a section to remove the
event listeners once Nym dies; I'll take a look at that tomorrow.