Three fixes, including issue #80
1) Fixed R4-E1 to only fire when an action has been taken.  Still too
much work to verify that the action was an ordnance action.

2) Boost AI's estimation of BTL-A4 Y-wing laser's to-hit probability,
which should force the BTL-A4 AI to always lead with the laser and get
both attacks.

3) Fix for a weird edge-case error most prevalent with R4-E1: Action
bombs dropped as the very last action of the phase did not activate
wrap_*("endaction") wrappers because they were ending via endnoaction.
I cannot see any reason this should be, since the bombs are dropped via
an action doaction rather than donoaction.  Fixed, and now R4-E1
reliably fires even as the last action of the turn.