Fix edge case where R4-E1 can't parse PTL action list due to PTL being bad
getactionbarlist(), alone of all get*barlist() functions, requires a
true or false value to determine if the check is being done as the
end-of-maneuver action or not.  This means that any wrapping of
getactionbarlist() has to expect two arguments: a true/false, and the
return value from getactionbarlist().

PTL breaks this by not passing in a true/false value.  It should,
however, always pass in "false" because to do othewise A) is not correct
to the function parameter requirement, and B) breaks wrapping.

This fixes that issue and makes R4-E1 work properly in all cases.