Bombardier has probably never worked correctly with Emon or Sol Sixxa
It looks like the original dev was not in the habit of following his own
specs.
If a function has parameter A, then any wrap_before or _after of that
function *must* include A as its first parameter, and the end result
(for wrap_after) as the sebsequent parameter.  In this case,
getbomblocation has a non-optional Bomb object parameter, so
wrap_after("getbomblocation") function definitions *must* take the Bomb
object as first parameter, and the actual list of locations as the
second parameter.