Start on making permalink loading work
team.js: added escape out if a team is converted to a squadlist before
its teamlist is set (such as when loading a permalink)

Also refactored parseASCII to call TeamList.inputASCII()

teamlist.js: made inputOldJuggler handle pilot strings with or without
stripped single quotes.  Copied parsing code from Teams now reports if
an upgrade is not found *or* not allowed for this pilot (this check may
be removed later though)

xwings.js: call to TEAMS[x].tosquadron(s) appear to be redundant.  Calls
to toJSON() definitely are.  Commented out for now.

Remaining issues: the x,y,rot coordinates encoded in the ASCII strings
are just discarded currently.  I either need to add a coordinate field
to simpleUnits or figure out a way to cheat that info in after parsing
the ASCII info.