Actual final Maul update, and fix for Uniques remaining disabled after changing faction in Create page
page_create.js: an exception to the faction restriction on a card really
only matters if the ship's faction and card's faction *don't* match;
otherwise, we should assume that any same-faction cards are allowed.
Adding this check makes Maul work within SCUM correctly.

team.js: Since we no longer rely on the "generics" global list for
populating text fields or generating squadrons for combat, there is no
reason to *not* clean it out when switching factions at squad creation
time.  But just deleting all entries is not enough; we have to unlock
any unique cards or pilots so that they are available in the next
faction.  This has been an ongoing but intermittent issue.